Add EasyEffects (PulseEffects) primary profile
Credit to part of this profile goes to Digitalone1 at:
https://github.com/Digitalone1/EasyEffects-Presets. Works
well on the Dell XPS 9520 internal speakers.

It has a volume normalizer which does a great job of keeping
audio at a consistent level on the internal speakers, an
EQ that widens the soundstage a bit, and adjustments for more
bass response since this machine has 2 woofer drivers enabled in
Pop! OS.